adempiereEAM#7 Error executing 40_AM_Module_Process.xml Reference to issue https://github.com/adempiere/adempiereEAM/issues/7
Avoid that Column "Calculate Cost" from table "AM_MaintenancePattern" is
deleted. The column is needed for calculating costs in window
"Maintenance Pattern".

ToDo: the column has to be added as field of the corresponding window.

ToDo: for windows Preventive Maintenance, Request Service Order and Work
Order the column  "Calculate Cost" and the corresponding field have to
be created, too.

With this commit, the 4 xml files can be executed.